+++ src/codemirror/components/LanguageSelect.tsx
@@ -4,24 +4,21 @@ import { Flexbox, MaterialFileTypeIcon, Select, Text } from '@lobehub/ui';
 import { cx } from 'antd-style';
 import { type FC, useMemo } from 'react';
 
-import { useTranslation } from '@/editor-kernel/react/useTranslation';
-
-import { MODES } from '../../lib/mode';
+import { LANGUAGES } from '../constants';
+import type { LanguageSelectProps } from '../types';
 import { styles } from './style';
 
-export interface LanguageSelectProps {
-  /** 语言变更回调 */
-  onLanguageChange: (value: string) => void;
-  /** 当前选中的语言 */
-  selectedLang: string;
-}
-
-export const LanguageSelect: FC<LanguageSelectProps> = ({ selectedLang, onLanguageChange }) => {
-  const t = useTranslation();
-  // 语言选项，使用 useMemo 优化性能
+export const LanguageSelect: FC<LanguageSelectProps> = ({
+  selectedLang,
+  onLanguageChange,
+  options,
+  labels,
+  className,
+}) => {
+  const modes = options ?? LANGUAGES;
   const languageOptions = useMemo(
     () =>
-      MODES.map((mode) => ({
+      modes.map((mode) => ({
         aliases: mode.ext || [],
         label: (
           <Flexbox align={'center'} gap={4} horizontal>
@@ -40,13 +37,13 @@ export const LanguageSelect: FC<LanguageSelectProps> = ({ selectedLang, onLangua
         title: mode.ext?.length ? mode.ext.map((ext) => `*.${ext}`).join(',') : `*.${mode.value}`,
         value: mode.value,
       })),
-    [],
+    [modes],
   );
 
   return (
     <Flexbox
       align={'center'}
-      className={'cm-language-select'}
+      className={cx('cm-language-select', className)}
       gap={4}
       horizontal
       onClick={(e) => e.stopPropagation()}
@@ -55,18 +52,15 @@ export const LanguageSelect: FC<LanguageSelectProps> = ({ selectedLang, onLangua
         className={cx(styles.container)}
         filterOption={(input, option) => {
           const lang: string = input.toLowerCase();
-          // 支持按值匹配
           if ((option?.value as string)?.toLowerCase().startsWith(lang)) return true;
-          // 支持按名称匹配
           if (String(option?.label).toLowerCase().includes(lang)) return true;
-          // 支持按扩展名匹配
           if (option?.aliases?.some((ext: string) => ext.toLowerCase().startsWith(lang)))
             return true;
           return false;
         }}
         onChange={onLanguageChange}
         options={languageOptions}
-        placeholder={t('codemirror.selectLanguage')}
+        placeholder={labels?.selectLanguage ?? 'Select language'}
         showSearch
         size="small"
         value={selectedLang}
