+++ src/codemirror/components/MoreOptions.tsx
@@ -5,22 +5,7 @@ import { Switch } from 'antd';
 import { MoreHorizontalIcon } from 'lucide-react';
 import { type FC, useCallback } from 'react';
 
-import { useTranslation } from '@/editor-kernel/react/useTranslation';
-
-export interface MoreOptionsProps {
-  /** 行号显示变更回调 */
-  onShowLineNumbersChange: (checked: boolean) => void;
-  /** Tab 大小变更回调 */
-  onTabSizeChange: (value: number | null) => void;
-  /** 制表符使用变更回调 */
-  onUseTabsChange: (checked: boolean) => void;
-  /** 是否显示行号 */
-  showLineNumbers: boolean;
-  /** Tab 大小 */
-  tabSize: number;
-  /** 是否使用制表符 */
-  useTabs: boolean;
-}
+import type { MoreOptionsProps } from '../types';
 
 export const MoreOptions: FC<MoreOptionsProps> = ({
   tabSize,
@@ -29,9 +14,8 @@ export const MoreOptions: FC<MoreOptionsProps> = ({
   onUseTabsChange,
   showLineNumbers,
   onShowLineNumbersChange,
+  labels,
 }) => {
-  const t = useTranslation();
-
   const handleTabSizeChange = useCallback(
     (value: number | null = 2) => {
       const v = value === null ? 2 : value;
@@ -46,7 +30,7 @@ export const MoreOptions: FC<MoreOptionsProps> = ({
       content={
         <Flexbox gap={8} style={{ minWidth: 240 }}>
           <Flexbox align={'center'} gap={8} horizontal justify={'space-between'}>
-            <Text>{t('codemirror.tabSize')}</Text>
+            <Text>{labels?.tabSize ?? 'Tab size'}</Text>
             <InputNumber
               max={8}
               min={1}
@@ -56,11 +40,11 @@ export const MoreOptions: FC<MoreOptionsProps> = ({
             />
           </Flexbox>
           <Flexbox align={'center'} gap={8} horizontal justify={'space-between'}>
-            <Text>{t('codemirror.useTabs')}</Text>
+            <Text>{labels?.useTabs ?? 'Use tabs'}</Text>
             <Switch checked={useTabs} onChange={onUseTabsChange} size="small" />
           </Flexbox>
           <Flexbox align={'center'} gap={8} horizontal justify={'space-between'}>
-            <Text>{t('codemirror.showLineNumbers')}</Text>
+            <Text>{labels?.showLineNumbers ?? 'Show line numbers'}</Text>
             <Switch checked={showLineNumbers} onChange={onShowLineNumbersChange} size="small" />
           </Flexbox>
         </Flexbox>
