+++ src/codemirror/components/Toolbar.tsx
@@ -4,24 +4,10 @@ import { ActionIcon, Flexbox } from '@lobehub/ui';
 import { ChevronDown, ChevronRight } from 'lucide-react';
 import { type FC } from 'react';
 
-import CopyButton from './CopyButton';
+import type { ToolbarProps } from '../types';
+import { CopyButton } from './CopyButton';
 import { LanguageSelect } from './LanguageSelect';
 import { MoreOptions } from './MoreOptions';
-
-export interface ToolbarProps {
-  expand?: boolean;
-  onClick?: () => void;
-  onCopy: () => void;
-  onLanguageChange: (value: string) => void;
-  onShowLineNumbersChange: (checked: boolean) => void;
-  onTabSizeChange: (value: number | null) => void;
-  onUseTabsChange: (checked: boolean) => void;
-  selectedLang: string;
-  showLineNumbers: boolean;
-  tabSize: number;
-  toggleExpand?: () => void;
-  useTabs: boolean;
-}
 
 export const Toolbar: FC<ToolbarProps> = ({
   selectedLang,
@@ -36,6 +22,8 @@ export const Toolbar: FC<ToolbarProps> = ({
   onClick,
   expand,
   toggleExpand,
+  labels,
+  languageOptions,
 }) => {
   return (
     <Flexbox
@@ -46,9 +34,15 @@ export const Toolbar: FC<ToolbarProps> = ({
       onClick={onClick}
       padding={4}
     >
-      <LanguageSelect onLanguageChange={onLanguageChange} selectedLang={selectedLang} />
+      <LanguageSelect
+        labels={labels}
+        onLanguageChange={onLanguageChange}
+        options={languageOptions}
+        selectedLang={selectedLang}
+      />
       <Flexbox gap={4} horizontal onClick={(e) => e.stopPropagation()}>
         <MoreOptions
+          labels={labels}
           onShowLineNumbersChange={onShowLineNumbersChange}
           onTabSizeChange={onTabSizeChange}
           onUseTabsChange={onUseTabsChange}
@@ -56,7 +50,7 @@ export const Toolbar: FC<ToolbarProps> = ({
           tabSize={tabSize}
           useTabs={useTabs}
         />
-        <CopyButton onCopy={onCopy} />
+        <CopyButton labels={labels} onCopy={onCopy} />
         <ActionIcon
           icon={expand ? ChevronDown : ChevronRight}
           onClick={toggleExpand}
