+++ tsconfig.json
@@ -19,6 +19,8 @@
       "@@/*": [".dumi/tmp/*"],
       "@lobehub/editor": ["./src"],
       "@lobehub/editor/*": ["./src/*"],
+      "@lobehub/editor/codemirror": ["./src/codemirror"],
+      "@lobehub/editor/codemirror/*": ["./src/codemirror/*"],
       "@lobehub/editor/headless": ["./src/headless"],
       "@lobehub/editor/headless/*": ["./src/headless/*"],
       "@lobehub/editor/react": ["./src/react"],
